Implemented KDE
Created a log function that returns 0 if the input value is 0. Use this instead of regular log. PDF function now has 2 modes. classic mode and log mode. calculate_likelihood similarly has classic and kde modes. The pose class is now modified to hold it's data. Training with the kde mode, causes the data to be stored in the pose objects.